Codeanalyzer/requirements.txt
tree-sitter==0.20.1
openai==1.3.0
python-dotenv==1.0.0
pytest==7.4.3